+++ android/app/build.gradle
@@ -31,6 +31,11 @@ android {
         disable 'InvalidPackage'
     }
 
+    dependencies {
+        implementation 'com.google.firebase:firebase-bom:17.0.0'
+        api 'com.google.firebase:firebase-core'
+    }
+
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
         applicationId "com.example.flyx"
@@ -64,6 +69,8 @@ dependencies {
     implementation 'com.google.firebase:firebase-database:16.1.0'
     implementation 'com.google.firebase:firebase-auth:16.2.0'
     implementation 'com.android.support:multidex:1.0.3'
+    implementation 'com.google.firebase:firebase-bom:17.0.0'
+    api 'com.google.firebase:firebase-core'
 
 }
 apply plugin: 'com.google.gms.google-services'
