+++ android/app/src/main/java/com/example/flyx/MainActivity.java
@@ -1,6 +1,7 @@
 package com.example.flyx;
 
 import android.os.Bundle;
+
 import io.flutter.app.FlutterActivity;
 import io.flutter.plugins.GeneratedPluginRegistrant;
