+++ android/settings.gradle
@@ -13,3 +13,5 @@ plugins.each { name, path ->
     include ":$name"
     project(":$name").projectDir = pluginDirectory
 }
+
+enableFeaturePreview('IMPROVED_POM_SUPPORT')
