--- lib/FloatingActionButton/fab_with_icons.dart
@@ -1,89 +1,0 @@
-import 'dart:math' as math;
-
-import 'package:flutter/material.dart';
-
-// https://stackoverflow.com/questions/46480221/flutter-floating-action-button-with-speed-dail
-class FabWithIcons extends StatefulWidget {
-  FabWithIcons({this.icons, this.onIconTapped});
-  final List<IconData> icons;
-  ValueChanged<int> onIconTapped;
-  @override
-  State createState() => FabWithIconsState();
-}
-
-class FabWithIconsState extends State<FabWithIcons>
-    with TickerProviderStateMixin {
-  AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 250),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      mainAxisSize: MainAxisSize.min,
-      children: List.generate(widget.icons.length, (int index) {
-        return _buildChild(index);
-      }).toList()
-        ..add(
-          _buildFab(),
-        ),
-    );
-  }
-
-  Widget _buildChild(int index) {
-    Color backgroundColor = Theme.of(context).cardColor;
-    Color foregroundColor = Theme.of(context).accentColor;
-    return Container(
-      height: 70.0,
-      width: 56.0,
-      alignment: FractionalOffset.topCenter,
-      child: ScaleTransition(
-        scale: CurvedAnimation(
-          parent: _controller,
-          curve: Interval(0.0, 1.0 - index / widget.icons.length / 2.0,
-              curve: Curves.easeOut),
-        ),
-        child: FloatingActionButton(
-          backgroundColor: backgroundColor,
-          mini: false,
-          shape: BeveledRectangleBorder(borderRadius: BorderRadius.circular(8)),
-          child: Icon(widget.icons[index], color: foregroundColor),
-          onPressed: () => _onTapped(index),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildFab() {
-    return Transform.rotate(
-      angle: -math.pi / 4,
-      child: FloatingActionButton(
-        onPressed: () {
-          if (_controller.isDismissed) {
-            _controller.forward();
-          } else {
-            _controller.reverse();
-          }
-        },
-        tooltip: 'pay',
-        child: Transform.rotate(
-            angle: -math.pi / -4, child: Icon(Icons.monetization_on)),
-        elevation: 2.0,
-        shape: BeveledRectangleBorder(borderRadius: BorderRadius.circular(8)),
-      ),
-    );
-  }
-
-  void _onTapped(int index) {
-    _controller.reverse();
-    widget.onIconTapped(index);
-  }
-}
