--- lib/FloatingActionButton/layout.dart
@@ -1,137 +1,0 @@
-// code from: https://github.com/matthew-carroll/flutter_ui_challenge_feature_discovery
-// TODO: Use https://github.com/matthew-carroll/fluttery/blob/master/lib/src/layout_overlays.dart
-import 'package:flutter/material.dart';
-
-class AnchoredOverlay extends StatelessWidget {
-  final bool showOverlay;
-  final Widget Function(BuildContext, Offset anchor) overlayBuilder;
-  final Widget child;
-
-  AnchoredOverlay({
-    this.showOverlay,
-    this.overlayBuilder,
-    this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return new Container(
-      child: new LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
-        return new OverlayBuilder(
-          showOverlay: showOverlay,
-          overlayBuilder: (BuildContext overlayContext) {
-            RenderBox box = context.findRenderObject() as RenderBox;
-            final center = box.size.center(box.localToGlobal(const Offset(0.0, 0.0)));
-
-            return overlayBuilder(overlayContext, center);
-          },
-          child: child,
-        );
-      }),
-    );
-  }
-}
-
-class OverlayBuilder extends StatefulWidget {
-  final bool showOverlay;
-  final Function(BuildContext) overlayBuilder;
-  final Widget child;
-
-  OverlayBuilder({
-    this.showOverlay = false,
-    this.overlayBuilder,
-    this.child,
-  });
-
-  @override
-  _OverlayBuilderState createState() => new _OverlayBuilderState();
-}
-
-class _OverlayBuilderState extends State<OverlayBuilder> {
-  OverlayEntry overlayEntry;
-
-  @override
-  void initState() {
-    super.initState();
-
-    if (widget.showOverlay) {
-      WidgetsBinding.instance.addPostFrameCallback((_) => showOverlay());
-    }
-  }
-
-  @override
-  void didUpdateWidget(OverlayBuilder oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    WidgetsBinding.instance.addPostFrameCallback((_) => syncWidgetAndOverlay());
-  }
-
-  @override
-  void reassemble() {
-    super.reassemble();
-    WidgetsBinding.instance.addPostFrameCallback((_) => syncWidgetAndOverlay());
-  }
-
-  @override
-  void dispose() {
-    if (isShowingOverlay()) {
-      hideOverlay();
-    }
-
-    super.dispose();
-  }
-
-  bool isShowingOverlay() => overlayEntry != null;
-
-  void showOverlay() {
-    overlayEntry = new OverlayEntry(
-      builder: widget.overlayBuilder,
-    );
-    addToOverlay(overlayEntry);
-  }
-
-  void addToOverlay(OverlayEntry entry) async {
-    print('addToOverlay');
-    Overlay.of(context).insert(entry);
-  }
-
-  void hideOverlay() {
-    print('hideOverlay');
-    overlayEntry.remove();
-    overlayEntry = null;
-  }
-
-  void syncWidgetAndOverlay() {
-    if (isShowingOverlay() && !widget.showOverlay) {
-      hideOverlay();
-    } else if (!isShowingOverlay() && widget.showOverlay) {
-      showOverlay();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return widget.child;
-  }
-}
-
-class CenterAbout extends StatelessWidget {
-  final Offset position;
-  final Widget child;
-
-  CenterAbout({
-    this.position,
-    this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return new Positioned(
-      top: position.dy,
-      left: position.dx,
-      child: new FractionalTranslation(
-        translation: const Offset(-0.5, -0.5),
-        child: child,
-      ),
-    );
-  }
-}
