+++ lib/style/clipper.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 class CustomShapeClipper extends CustomClipper<Path> {
@@ -7,11 +6,11 @@ class CustomShapeClipper extends CustomClipper<Path> {
     final Path path = Path();
     path.lineTo(0.0, size.height);
 
-    var firstEndPoint = Offset(size.width , size.height);
-    var firstControlpoint = Offset(size.width * 0.5, size.height*.80);
-    path.quadraticBezierTo(firstControlpoint.dx, firstControlpoint.dy, firstEndPoint.dx, firstEndPoint.dy);
-    
-  
+    var firstEndPoint = Offset(size.width, size.height);
+    var firstControlpoint = Offset(size.width * 0.5, size.height * .80);
+    path.quadraticBezierTo(firstControlpoint.dx, firstControlpoint.dy,
+        firstEndPoint.dx, firstEndPoint.dy);
+
     path.lineTo(size.width, 0.0);
     path.close();
     return path;
