+++ pubspec.yaml
@@ -19,21 +19,24 @@ environment:
 dependencies:
   bubble_bottom_bar: ^1.1.0
   transparent_image: ^0.1.0
-  google_sign_in: ^4.0.1+1
-  firebase_auth: ^0.8.1+4
-  firebase_core: ^0.3.1+1
+  google_sign_in: ^4.0.1+3
+  firebase_auth: ^0.8.3
+  firebase_core: ^0.3.2
   firebase_database: ^2.0.2
   local_auth: ^0.4.0+1
-  google_maps_flutter: ^0.3.0+2
+  google_maps_flutter: ^0.5.2
   date_range_picker: ^1.0.5
   http: ^0.12.0+1
-  cloud_firestore: ^0.9.5+2
+  cloud_firestore: ^0.9.11
   auto_size_text: ^1.1.0
   flutter_backdrop: ^1.0.3
   flip_card: ^0.2.1
   url_launcher: ^5.0.2
   groovin_widgets: ^1.3.2
-  rubber: ^0.2.7
+  rubber:
+    git:
+      url: https://github.com/mcrovero/rubber.git
+      ref: master
   rounded_modal: ^1.0.0
   geohash: ^0.2.1
   search_widget: ^0.1.0
@@ -81,15 +84,20 @@ flutter:
   fonts:
       - family: OpenSans
         fonts:
-         - asset: lib/assets/fonts/OpenSans-Bold.ttf
-         - asset: lib/assets/fonts/OpenSans-Regular.ttf
-           weight: 700
-           
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
+          - asset: lib/assets/fonts/OpenSans-Bold.ttf
+          - asset: lib/assets/fonts/OpenSans-BoldItalic.ttf
+          - asset: lib/assets/fonts/OpenSans-ExtraBold.ttf
+          - asset: lib/assets/fonts/OpenSans-ExtraBoldItalic.ttf
+          - asset: lib/assets/fonts/OpenSans-Italic.ttf
+          - asset: lib/assets/fonts/OpenSans-Light.ttf
+          - asset: lib/assets/fonts/OpenSans-LightItalic.ttf
+          - asset: lib/assets/fonts/OpenSans-Regular.ttf
+          - asset: lib/assets/fonts/OpenSans-SemiBold.ttf           
+          - asset: lib/assets/fonts/OpenSans-SemiBoldItalic.ttf           
+      - family: Dokdo
+        fonts:
+          - asset: lib/assets/fonts/Dokdo-Regular.ttf
+            
+    
   # For details regarding fonts from package dependencies,
   # see https://flutter.io/custom-fonts/#from-packages
